Restructure into local and global kernels
