+++ nfmc/algorithms/mh/local/mala.py
@@ -1,6 +1,6 @@
 from typing import Tuple, Union
 import torch
-from nfmc.algorithms.mh.local import LocalMHKernel
+from nfmc.algorithms.mh.local.base import LocalMHKernel
 from nfmc.util import compute_divergence_mask, grad_f, metropolis_acceptance_log_ratio, sum_except_batch, diag_mult
